# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel",
        "version": "1.0.0"
    },
    "paths": {
        "/api/blogs": {
            "post": {
                "tags": [
                    "Blogs"
                ],
                "summary": "Create a Blog",
                "description": "Create a new Blog.",
                "operationId": "163c9d485fea7b09d51b6ceb5b8e66fe",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BlogRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Blog created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/blogs/{id}": {
            "put": {
                "tags": [
                    "Blogs"
                ],
                "summary": "Update a Blog",
                "description": "Update an existing Blog.",
                "operationId": "3ea95b7e4df029108300c7b1cdaab4de",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BlogRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Blog updated successfully"
                    },
                    "404": {
                        "description": "Blog not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Blogs"
                ],
                "summary": "Delete a Blog",
                "description": "Delete a Blog.",
                "operationId": "0eff8bd8dd339995b4d69ea9abed49e7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Blog deleted successfully"
                    },
                    "404": {
                        "description": "Blog not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/blogs/search": {
            "get": {
                "tags": [
                    "Blogs"
                ],
                "summary": "Search for blogs",
                "description": "Search for Blogs.",
                "operationId": "2283b60aa636448cce2a854ea022130f",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search keyword",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "maxLength": 255
                        }
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Category ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "source",
                        "in": "query",
                        "description": "News source",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "maxLength": 255
                        }
                    },
                    {
                        "name": "origin",
                        "in": "query",
                        "description": "Origin of the news",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "NYT",
                                "Guardian",
                                "NewsAPI"
                            ]
                        }
                    },
                    {
                        "name": "from",
                        "in": "query",
                        "description": "Start date for filtering",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2024-02-25",
                            "nullable": true
                        }
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "description": "End date for filtering",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2024-02-25",
                            "nullable": true
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Per page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Search results returned successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "BlogRequest": {
                "required": [
                    "title",
                    "content",
                    "url",
                    "source",
                    "authors"
                ],
                "properties": {
                    "title": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "New blog"
                    },
                    "snippet": {
                        "type": "string",
                        "maxLength": 500,
                        "example": "New blog content...",
                        "nullable": true
                    },
                    "content": {
                        "type": "string",
                        "example": "New blog content ..."
                    },
                    "url": {
                        "type": "string",
                        "format": "url",
                        "maxLength": 500,
                        "example": "https://example.com/blog/new-blog"
                    },
                    "published_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-02-25T12:00:00Z",
                        "nullable": true
                    },
                    "source": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "The New York Times"
                    },
                    "origin": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "NYT",
                        "nullable": true
                    },
                    "authors": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "maxLength": 255,
                            "example": "Shayan"
                        }
                    },
                    "image": {
                        "type": "string",
                        "format": "url",
                        "maxLength": 500,
                        "example": "https://example.com/images/1.jpg",
                        "nullable": true
                    },
                    "categories": {
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "example": 1
                        },
                        "nullable": true
                    }
                },
                "type": "object"
            }
        },
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}